[WIP] Add Event type
